# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateSpecializationRequest:
    properties:
      code:
        maxLength: 20
        minLength: 1
        type: string
      fhir_code:
        maxLength: 50
        type: string
      fhir_system:
        maxLength: 200
        type: string
      gelar:
        maxLength: 20
        type: string
      is_aktif:
        type: boolean
      job_title_id:
        type: integer
      kategori_id:
        type: integer
      label:
        maxLength: 150
        minLength: 1
        type: string
      lama_pendidikan_tahun:
        maximum: 20
        minimum: 0
        type: integer
    required:
    - code
    - label
    type: object
...
  dto.JobTitleSimpelResponse:
    properties:
      code:
        type: string
      id:
        type: integer
      label:
        type: string
    type: object
...
  dto.SpecializationKategoriSimpelResponse:
    properties:
      code:
        type: string
      fhir_code:
        type: string
      id:
        type: integer
      label:
        type: string
    type: object
  dto.SpecializationResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/httputil.UserData'
      fhir_code:
        type: string
      fhir_system:
        type: string
      gelar:
        type: string
      id:
        type: integer
      is_aktif:
        type: boolean
      job_title:
        $ref: '#/definitions/dto.JobTitleSimpelResponse'
      kategori:
        $ref: '#/definitions/dto.SpecializationKategoriSimpelResponse'
      label:
        type: string
      lama_pendidikan_tahun:
        type: integer
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/httputil.UserData'
    type: object
  dto.SpecializationSelectResponse:
    properties:
      code:
        type: string
      gelar:
        type: string
      id:
        type: integer
      label:
        type: string
    type: object
...
  dto.UpdateSpecializationRequest:
    properties:
      code:
        maxLength: 20
        minLength: 1
        type: string
      fhir_code:
        maxLength: 50
        type: string
      fhir_system:
        maxLength: 200
        type: string
      gelar:
        maxLength: 20
        type: string
      is_aktif:
        type: boolean
      job_title_id:
        type: integer
      kategori_id:
        type: integer
      label:
        maxLength: 150
        minLength: 1
        type: string
      lama_pendidikan_tahun:
        maximum: 20
        minimum: 0
        type: integer
    type: object
...
  httputil.UserData:
    properties:
      id:
        type: integer
      name:
        type: string
      username:
        type: string
    type: object
...
  response.MyGoResponse:
    description: Standard JSON response
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
host: localhost:1323
info:
  contact:
    email: support@neosim.id
    name: Neosim Support
  description: REST API untuk Neosim — sistem manajemen informasi
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Neosim API V1
  version: "1.0"
paths:
...
  /kepegawaian/jabatan/specializations:
    get:
      consumes:
      - application/json
      description: Get paginated list of Specialization
      parameters:
      - description: Filter by code (partial match)
        in: query
        name: code
        type: string
      - description: Filter by label (partial match)
        in: query
        name: label
        type: string
      - description: Filter by job title ID
        in: query
        name: job_title_id
        type: integer
      - description: Filter by kategori ID
        in: query
        name: kategori_id
        type: integer
      - description: Filter by status aktif
        in: query
        name: is_aktif
        type: boolean
      - description: Sort column (code,label,job_title_id,kategori_id,gelar,lama_pendidikan_tahun,fhir_code,
          created_at, updated_at)
        enum:
        - code
        - label
        - job_title_id
        - kategori_id
        - gelar
        - lama_pendidikan_tahun
        - fhir_code
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SpecializationResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get list of Specialization
      tags:
      - kepegawaian/jabatan/specializations
    post:
      consumes:
      - application/json
      description: Create New Specialization
      parameters:
      - description: Create Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSpecializationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SpecializationResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create Specialization
      tags:
      - kepegawaian/jabatan/specializations
  /kepegawaian/jabatan/specializations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Specialization by :id
      parameters:
      - description: Specialization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MyGoResponse'
      security:
      - BearerAuth: []
      summary: Delete Specialization
      tags:
      - kepegawaian/jabatan/specializations
    get:
      consumes:
      - application/json
      description: Get Specialization by :id
      parameters:
      - description: Specialization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SpecializationResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get Specialization
      tags:
      - kepegawaian/jabatan/specializations
    put:
      consumes:
      - application/json
      description: Update Specialization by :id
      parameters:
      - description: Specialization ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSpecializationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SpecializationResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update Specialization
      tags:
      - kepegawaian/jabatan/specializations
  /kepegawaian/jabatan/specializations/select:
    get:
      consumes:
      - application/json
      description: Get list Specialization aktif untuk dropdown/select (tanpa pagination)
      parameters:
      - description: Search by label, code, atau gelar
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SpecializationSelectResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get select list of Specialization
      tags:
      - kepegawaian/jabatan/specializations
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"